+++ tuning_fork_kino/test/kino_tuning_fork/live_loops_cell_test.exs
@@ -30,6 +30,7 @@ defmodule KinoTuningFork.LiveLoopsCellTest do
   defp cell(loops) do
     {kino, _source} = start_smart_cell!(LiveLoopsCell, %{"loops" => loops})
     _data = connect(kino)
+    push_event(kino, "listening", %{"on" => true})
 
     kino
   end