+++ tuning_fork_kino/test/kino_tuning_fork/live_patterns_cell_test.exs
@@ -17,6 +17,7 @@ defmodule KinoTuningFork.LivePatternsCellTest do
   defp cell(attrs \\ %{}) do
     {kino, _source} = start_smart_cell!(LivePatternsCell, attrs)
     _data = connect(kino)
+    push_event(kino, "listening", %{"on" => true})
 
     kino
   end