+++ tuning_fork_kino/test/kino_tuning_fork/midi_monitor_cell_test.exs
@@ -44,7 +44,10 @@ defmodule KinoTuningFork.MidiMonitorCellTest do
       playing: false,
       cps: 0.5,
       clock: false,
-      events: [{:out, :start, 0} | for(n <- 1..20, do: {:in, {:note_on, 0, 40 + n, 100}, n})]
+      events: [
+        {:out, :start, 0},
+        {:out, :clock, 0} | for(n <- 1..20, do: {:in, {:note_on, 0, 40 + n, 100}, n})
+      ]
     }
 
     shown = MidiMonitorCell.page_state(state)
@@ -54,6 +57,7 @@ defmodule KinoTuningFork.MidiMonitorCellTest do
     refute Map.has_key?(shown, :voice)
     assert length(shown.events) == 12
     assert Enum.take(shown.events, 2) == [[:out, [:start], 0], [:in, [:note_on, 0, 41, 100], 1]]
+    refute Enum.any?(shown.events, &(&1 == [:out, [:clock], 0]))
     assert {:ok, _json} = Jason.encode(shown)
 
     assert MidiMonitorCell.page_state(%{state | input: nil, output: nil, events: []}).input == nil